[CHORE] #104 : 모듈 analyze -> analysis rename
