+++ android/feature/analysis/src/main/AndroidManifest.xml
