+++ android/feature/analysis/src/main/res/layout/dialog_choose_tools.xml
