+++ android/feature/analysis/src/main/res/layout/fragment_camera.xml
