+++ android/feature/analysis/src/main/res/layout/fragment_home.xml
@@ -5,7 +5,7 @@
     <data>
         <variable
             name="viewModel"
-            type="com.jaino.analyze.home.HomeViewModel" />
+            type="com.jaino.analysis.home.HomeViewModel" />
     </data>
     <androidx.constraintlayout.widget.ConstraintLayout
         android:layout_width="match_parent"
